pyproject.toml
[tool.poetry]
name = "sashimi"
version = "0.1.0"
description = "Exact search engine"
authors = ["Yaroslav <yaroslaff@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
pyyaml = "^6.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.15"
evalidate = "^2.0.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
